+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Method `fromArray` and `fromConfig` to graphql client.
+- A `SessionUser` to use for web auth instead of `GenericUser`.
 
 ### Changed
 - **BREAKING**: Do not use middleware aliases.
+- Renamed session key for web auth.
 
 ## [0.22.3] - 2023-05-22